Added support for loading Stats file from URL
In production, it is generally favourable to have the stats file loaded
from some external URL (and not from filesystem). This is especially
useful for heroku where we no longer need to build the static assets
every time we deploy.

In this commit, I have added an implementation for specifying
`STATS_FILE_URL` in the configuration. Also added corresponding tests
with a mocked request. (It could use a few more test cases to handle the
exceptions though.)

-- Prashant