Add a setting to configure skip common chunks behaviour globally
You can now specify a 'SKIP_COMMON_CHUNKS' setting globally for each
Webpack configuration. This will perform the same 'skip_common_chunks'
logic from the 'render_bundle' template tag without having to add
a parameter to every tag.

To maintain backwards compatibility we default to 'False' and allow
common chunks to be rendered to the HTML document as script tags. In the
future, it will be easy to change this default behaviour if desired.

I've added test coverage to ensure that the global setting can be
override by the existing template tag parameter and that we maintain
backwards compatibility.